refactor: semantics and organization
